+++ mrmustard/math/lattice/strategies/compactFock/julia/compactFock_1leftoverMode_amps.jl
