+++ mrmustard/math/lattice/strategies/compactFock/julia/compactFock_1leftoverMode_grad.jl
