+++ mrmustard/math/lattice/strategies/compactFock/julia/compactFock_diagonal_amps.jl
