+++ mrmustard/math/lattice/strategies/compactFock/julia/compactFock_helperFunctions.jl
