+++ mrmustard/math/lattice/strategies/compactFock/python/compactFock_1leftoverMode_amps.py
@@ -8,7 +8,7 @@ import numpy as np
 from numba import int64, njit
 from numba.cpython.unsafe.tuple import tuple_setitem
 
-from mrmustard.math.lattice.strategies.compactFock.compactFock_helperFunctions import (
+from mrmustard.math.lattice.strategies.compactFock.python.compactFock_helperFunctions import (
     SQRT,
     construct_dict_params,
     repeat_twice,
